backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-docx==1.1.2
anthropic==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.9
